\name{aihrio}
\alias{aihrio}
\docType{data}
\title{Sample dataset}
\description{This is a dataset for Poisson-Gamma Additive Models functions testing.
}
\usage{data(aihrio)}
\format{
A data frame with 365 observations on the following 33 variables.
  \describe{
    \item{DATE}{a factor with levels}
    \item{TIME}{a numeric vector}
    \item{ITRESP65}{a numeric vector}
    \item{ITCIRC65}{a numeric vector}
    \item{ITDPOC65}{a numeric vector}
    \item{ITPNM65}{a numeric vector}
    \item{ITAVC65}{a numeric vector}
    \item{ITIAM65}{a numeric vector}
    \item{ITDIC65}{a numeric vector}
    \item{ITTCA65}{a numeric vector}
    \item{ITRESP5}{a numeric vector}
    \item{ITPNEU5}{a numeric vector}
    \item{ITDPC5}{a numeric vector}
    \item{WEEK}{a numeric vector}
    \item{MON}{a numeric vector}
    \item{TUE}{a numeric vector}
    \item{WED}{a numeric vector}
    \item{THU}{a numeric vector}
    \item{FRI}{a numeric vector}
    \item{SAT}{a numeric vector}
    \item{SUN}{a numeric vector}
    \item{HOLIDAYS}{a numeric vector}
    \item{MONTH}{a numeric vector}
    \item{warm.season}{a numeric vector}
    \item{tmpmed}{a numeric vector}
    \item{tmpmin}{a numeric vector}
    \item{tmpmax}{a numeric vector}
    \item{wet}{a numeric vector}
    \item{rain}{a numeric vector}
    \item{rainy}{a numeric vector}
    \item{PM}{a numeric vector}
    \item{SO2}{a numeric vector}
    \item{CO}{a numeric vector}
  }
}
\details{This is a reduced dataset of those used to estimate possible effects of air pollution on hospital admissions outcomes in Universidade do Estado do Rio de Janeiro, Rio de Janeiro, Brasil.
}
\source{
Secretary for the Environment of the Rio de Janeiro City, Brazilian Ministry of Defense and Brazilian Ministry of Health
}
\author{Washington Leite Junger \email{wjunger@ims.uerj.br} and Antonio Ponce de Leon \email{ponce@ims.uerj.br}
}
\keyword{datasets}
