\name{elapsedtime}
\alias{elapsedtime}
\title{Utility function}
\description{Gadget to compute the elapsed time of a process
}
\usage{
elapsedtime(st, et)
}
\arguments{
  \item{st}{start time}
  \item{et}{end time}
}
\details{Start and end times can be obtained with \code{\link{proc.time}}.
}
\value{String with the elapsed time in \code{hh:mm:ss} format.
}
\author{Washington Leite Junger \cr \email{wjunger@ims.uerj.br}
}
\seealso{\code{\link{pgam}}}
	\keyword{ts}
	\keyword{smooth}
	\keyword{regression}
