\name{envelope}
\alias{envelope}
\title{Generic function for simulated envelope generation.
}
\description{A normal plot with simulated envelope of the residual is produced.
}
\usage{
envelope(object, ...)
}
\arguments{
    \item{object}{object holding the fitted model}
	\item{\dots}{further arguments to passed to methods}
}
}
\value{An object of class \code{envelope} holding the information needed to plot the envelope.
}
\references{
Atkinson, A. C. (1985) Plots, transformations and regression : an introduction to graphical methods of diagnostic regression analysis. Oxford Science Publications, Oxford.
}
\author{Washington Leite Junger \email{wjunger@ims.uerj.br}
}
	\keyword{ts}
	\keyword{smooth}
	\keyword{regression}
