\name{fnz}
\alias{fnz}
\title{Utility function}
\description{Return the index of first non-zero observation of a variable.
}
\usage{
fnz(y)
}
\arguments{
  \item{y}{variable vector}
}
\value{The index of first non-zero value.
}
\author{Washington Leite Junger \cr \email{wjunger@ims.uerj.br}
}
\note{This function is not intended to be called directly.
}
\seealso{\code{\link{pgam}}}
	\keyword{ts}
	\keyword{smooth}
	\keyword{regression}
