\name{print.pgam}
\alias{print.pgam}
\title{Model output}
\description{Print model information}
\usage{
\method{print}{pgam}(x, digits, ...)
}
\arguments{
  \item{x}{object of class \code{summary.pgam} holding the fitted model information}
  \item{digits}{number of decimal places for output}
  \item{\dots}{further arguments passed to method}
}
\details{This function only prints out the information. 
}
\value{No value is returned.
}
\author{Washington Leite Junger \email{wjunger@ims.uerj.br} and Antonio Ponce de Leon \email{ponce@ims.uerj.br}
}
\seealso{\code{\link{pgam}}, \code{\link{predict.pgam}}}
	\keyword{ts}
	\keyword{smooth}
	\keyword{regression}
